+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # QOwnNotes Web Companion Changelog
 
 ## 19.1.11
+- added support to show a success message from QOwnNotes if bookmarks were created
+    - you need QOwnNotes 19.1.9 or higher for this feature
 - added more translation
 
 ## 19.1.10